Add filtered paths, that are not owned after building
This simplifies a number of common scenarios where the four-step process
of setup, register, build, retrieve is far too involved. Instead, we
pass a mutable reference to the buffer that we would like have
constructed as a local variable—wrapped in a single enum item that
denotes the interpretation of said Path—and call build in a single call
chain.