Get rid of distinction of File/Tree in filter
This was meant to be a small, convenience API for simple cases. In the
end a tree and a path will get the same treatment in git. It's far
easier if you don't have to do any wrapping yourself.

At least my personal experimentation shows that the extra import and
call to wrap might be interesting in a type-safety perspective but is of
little to no relevance in practice. So it's just a hassle.