Simplify registered paths to Files
Same justification as for rewrite, there is no inherent reason to
differentiate a single file from a tree. The interface is exactly the
same for both of them: path in—path out.

There may at a later date be justification for other types (hello, glob
paths) but even then it's questionable if we need a separate return
value. We might just need more internal state. At the end of the day
the caller expects a path to find their files under.