Switch to fs2 for file locks
This seems to fix a dead-lock situation with the other library, Maybe
suppose its use of CondVars for signalling is not quite as it should be
and misses some events? In any case that library is much more popular
and probably stable (I had initially discarded it due to maintenance
issues and its CI failing but you know what they say about perfect code
not needing more evolution).