Fix unstable git sparse-checkout set
Changed in some version around git 2.73.2, pattern specifications are
allowed only with an explicit `--no-cone` argument. As we make use of
those for precise file lists this is now our default. It's not entirely
clear how to unify with the advantages of using cone/sparse index, we'll
have to review if all patterns can also be applied as a full cone filter
without risk of data corruption.

One spurious issue was, I believe, that quoting had previously not been
supported but now "C-style quoted strings" are allowable for path names.